requirements.txt
amqp==2.2.2
billiard==3.5.0.3
boto==2.48.0
boto3==1.6.1
botocore==1.9.1
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.9
django-autocomplete-light==3.2.10
django-celery-results==1.0.1
django-recaptcha2==1.0.3
django-storages==1.6.5
djangorestframework==3.7.7
docutils==0.14
futures==3.2.0
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
kombu==4.1.0
MySQL-python==1.2.5
Pillow==5.0.0
python-dateutil==2.6.1
pytz==2017.3
reportlab==3.4.0
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
urllib3==1.22
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
